# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kabadiwala Connect API",
    "description": "Vernacular, low-literacy offline-tolerant platform connecting informal e-waste collectors with authorized recyclers.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Request OTP for collector phone login",
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify OTP and obtain access token",
        "operationId": "verify_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/pin/setup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Setup 4-digit PIN for collector",
        "operationId": "setup_pin_api_v1_auth_pin_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PINSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/pin/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login collector using 4-digit PIN",
        "operationId": "login_with_pin_api_v1_auth_pin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PINLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login recycler or admin using email and password",
        "operationId": "login_email_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register recycler or admin account",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current authenticated user profile",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/ledger/entries": {
      "get": {
        "tags": [
          "Collector Ledger"
        ],
        "summary": "List collector ledger entries",
        "operationId": "list_ledger_entries_api_v1_ledger_entries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerEntryResponse"
                  },
                  "title": "Response List Ledger Entries Api V1 Ledger Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Collector Ledger"
        ],
        "summary": "Record a cash receipt or transaction entry in ledger",
        "operationId": "create_ledger_entry_api_v1_ledger_entries_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedgerEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/summary": {
      "get": {
        "tags": [
          "Collector Ledger"
        ],
        "summary": "Get collector ledger balance summary",
        "operationId": "get_ledger_summary_api_v1_ledger_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "EmailLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "EmailLoginRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LedgerEntryCreate": {
        "properties": {
          "client_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Entry Id",
            "description": "Offline client UUID for idempotency"
          },
          "lot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Id"
          },
          "entry_type": {
            "type": "string",
            "pattern": "^(credit|debit)$",
            "title": "Entry Type"
          },
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount"
          },
          "payment_mode": {
            "type": "string",
            "pattern": "^(cash_received|pending|digital_paid)$",
            "title": "Payment Mode",
            "default": "cash_received"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description"
          },
          "recorded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          }
        },
        "type": "object",
        "required": [
          "entry_type",
          "amount",
          "description"
        ],
        "title": "LedgerEntryCreate"
      },
      "LedgerEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "collector_id": {
            "type": "string",
            "title": "Collector Id"
          },
          "lot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Id"
          },
          "entry_type": {
            "type": "string",
            "title": "Entry Type"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "payment_mode": {
            "type": "string",
            "title": "Payment Mode"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "balance_after": {
            "type": "number",
            "title": "Balance After"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "collector_id",
          "lot_id",
          "entry_type",
          "amount",
          "payment_mode",
          "description",
          "balance_after",
          "recorded_at",
          "created_at"
        ],
        "title": "LedgerEntryResponse"
      },
      "LedgerSummary": {
        "properties": {
          "collector_id": {
            "type": "string",
            "title": "Collector Id"
          },
          "total_earned_inr": {
            "type": "number",
            "title": "Total Earned Inr"
          },
          "cash_received_inr": {
            "type": "number",
            "title": "Cash Received Inr"
          },
          "pending_inr": {
            "type": "number",
            "title": "Pending Inr"
          },
          "current_balance_inr": {
            "type": "number",
            "title": "Current Balance Inr"
          },
          "total_transactions": {
            "type": "integer",
            "title": "Total Transactions"
          }
        },
        "type": "object",
        "required": [
          "collector_id",
          "total_earned_inr",
          "cash_received_inr",
          "pending_inr",
          "current_balance_inr",
          "total_transactions"
        ],
        "title": "LedgerSummary"
      },
...
      "OTPRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{9,14}$",
            "title": "Phone",
            "description": "E.164 phone number"
          },
          "language": {
            "type": "string",
            "pattern": "^(mr|hi|en)$",
            "title": "Language",
            "description": "Preferred language",
            "default": "mr"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "OTPRequest"
      },
      "OTPRequestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "otp_sent"
          },
          "retry_after_seconds": {
            "type": "integer",
            "title": "Retry After Seconds",
            "default": 60
          },
          "mock_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock Otp"
          }
        },
        "type": "object",
        "title": "OTPRequestResponse"
      },
      "OTPVerifyRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{9,14}$",
            "title": "Phone"
          },
          "otp": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Otp"
          },
          "operating_area": {
            "type": "string",
            "title": "Operating Area",
            "description": "Operating district",
            "default": "Mumbai Suburban"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "otp"
        ],
        "title": "OTPVerifyRequest"
      },
      "PINLoginRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{9,14}$",
            "title": "Phone"
          },
          "pin": {
            "type": "string",
            "pattern": "^\\d{4}$",
            "title": "Pin"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "pin"
        ],
        "title": "PINLoginRequest"
      },
      "PINSetupRequest": {
        "properties": {
          "pin": {
            "type": "string",
            "pattern": "^\\d{4}$",
            "title": "Pin",
            "description": "4-digit numeric PIN"
          }
        },
        "type": "object",
        "required": [
          "pin"
        ],
        "title": "PINSetupRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "collector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Id"
          },
          "has_pin": {
            "type": "boolean",
            "title": "Has Pin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role"
        ],
        "title": "TokenResponse"
      },
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "collector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Id"
          },
          "recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recycler Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserProfileResponse"
      },
      "UserRegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "pattern": "^(recycler|admin)$",
            "title": "Role",
            "default": "recycler"
          },
          "recycler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recycler Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserRegisterRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}